move scopes to trait
