+++ app/Models/Traits/HasAllergens.php
@@ -3,6 +3,7 @@
 namespace App\Models\Traits;
 
 use App\Models\Allergen;
+use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 
 trait HasAllergens
@@ -12,13 +13,22 @@ trait HasAllergens
         return $this->belongsToMany(Allergen::class);
     }
 
-    protected function getIsVeganAttribute(): bool
+    protected function scopeGlutenFree(Builder $query): Builder
     {
-        return $this->allergens()->where('is_animal_product', false)->count() > 0;
+        return $query->withoutAllergens(Allergen::firstWhere('name', 'like', '%gluten%'));
     }
 
-    protected function getIsVegetarian(): bool
+    protected function scopeWithoutAllergens(Builder $query, Allergen ...$allergens): Builder
     {
-        return $this->allergens()->where('is_animal_product', false)->count() > 0;
+        return $query->whereDoesntHave('allergens', function (Builder $query) use ($allergens) {
+            return $query->whereIn('allergens.id', collect($allergens)->pluck('id'));
+        });
+    }
+
+    protected function scopeWithAllergens(Builder $query, Allergen ...$allergens): Builder
+    {
+        return $query->whereHas('allergens', function (Builder $query) use ($allergens) {
+            return $query->whereIn('allergens.id', collect($allergens)->pluck('id'));
+        });
     }
 }
